# [REF] models, fields: refactor fields_get