@beledouxdenis 2022-03-31
> While looking for calls to `fields_get()`, I noticed that many calls can be optimized to retrieve only what they need.

Indeed, that's the main point for this change